git-grep-blame: handle case where git-grep matches binary file
In the case where a binary file contains what we are trying to git-grep,
the regular expression that tries to match the line will fail. We then
try to call the `each` method on a nil object, which results in an
exception.

This patch checks if there was a match between the regular expression
and the line of git-grep output before carrying on.

Signed-off-by: Pang Yan Han <pangyanhan@gmail.com>