git-grep-filter: Fit output within terminal columns
To this end, we introduce a new function, colorized_line_truncate_to_fit ,
which computes the longest prefix of a `grep` colorized line that can fit
within the number of columns in the terminal. This method ignores the
length of color escape characters from `grep` and substrings the original
colorized line accordingly.

With this change, git-grep-filter should give a similar output to `git grep -n`

Signed-off-by: Yan Han Pang <539739+yanhan@users.noreply.github.com>