Convert git-grep-filter to use Haskell & Shelly.hs
Have always wanted to use Haskell for practical purposes for a while, but
it's seems so difficult to do so. Shelly.hs comes to the rescue here.

In this patch, git-grep-filter is ported to use Haskell, mainly using the
Shelly.hs library (https://github.com/yesodweb/Shelly.hs).

The Ruby git-grep-filter will remain for historical purposes.

Signed-off-by: Yan Han Pang <539739+yanhan@users.noreply.github.com>