GitUtils.truncated_colorized_line_io - Regex truncation
This function makes use of the Data.Text.ICU.Regex module and the IO monad.
Although we don't make use of the GitUtils.truncated_colorized_line_io in
our code, it serves as an example of how to use the module and
regular expressions in Haskell.

Its code is also pretty messy due to hacking it out under heavy
time constraint.

Signed-off-by: Yan Han Pang <539739+yanhan@users.noreply.github.com>