git-grep-filter - dont colorize patterns in filename prefix
`git grep -n` output lines in the following format:

    filename:linenumber: matchingLine

When the pattern supplied is found in the `filename:linenumber:` prefix,
they will be highlighted in `less -R`.

This patch addresses the issue by stripping off the `filename:linenumber:`
prefix before outputting to a temporary file, so that the `grep` program
does not run on the prefixes. On a separate step, the prefix is then
concatenated with the colorized grep text before outputting to another
temporary file that is supplied to `less -R` .

Signed-off-by: Pang Yan Han <pangyanhan@gmail.com>